docs: add sample topologies

+++ samples/standalone/standalone.sh
@@ -1,23 +1,20 @@
 #!/usr/bin/env bash
 
-# Basic config.
-port=27017
-
-# Rebuild data directories.
-if [[ -d './data' ]]; then
-    rm -rf ./data
-fi
+# Rebuild data and log directories.
+if [[ -d './data' ]]; then rm -rf ./data; fi
 mkdir ./data
+if [[ -d './log' ]]; then rm -rf ./log; fi
+mkdir ./log
 
 # Start the standalone process.
-mongod --fork --logpath ./standalone.log --smallfiles --oplogSize 50 --port 27017 --dbpath ./data 
+mongod --fork --logpath ./log/standalone.log --smallfiles --oplogSize 50 --port 27017 --dbpath ./data 
 
 # Wait a bit of time for the process to start.
 sleep 10
 
 # Show the process info.
 echo
-echo "Standalone MongoDB Running on ${port}"
+echo "Standalone MongoDB Running on 27107"
 echo
 echo "Monitor with: "
-echo "  mongo-monitor localhost:${port}"
+echo "  mongo-monitor localhost:27017"
